test.py
import os
from dotenv import load_dotenv

from langchain_neo4j import Neo4jGraph

load_dotenv()
graph = Neo4jGraph(
    url=os.getenv("NEO4J_URI"),
    username=os.getenv("NEO4J_USERNAME"),
    password=os.getenv("NEO4J_PASSWORD"),
)


def main():
    result = graph.query("MATCH(n) RETURN COUNT(n) as node_count")
    print(result[0])


if __name__ == "__main__":
    main()