pyproject.toml
[project]
name = "neo4j-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==0.3.23",
    "langchain-google-genai==2.1.3",
    "langchain-neo4j==0.3.0",
    "langchain-openai==0.3.13",
    "numpy==1.26.4",
    "python-dotenv>=1.2.2",
]
